-----Changelog----- -Corregidas las texturas (la imagen tiene que estar invertida o bien, asignar  1 - u & 1 - v) -Falta implementación de 3era persona

+++ TextureVertex.cpp
@@ -28,7 +28,7 @@ GLfloat TextureVertex::getV()
 	return V;
 }
 
-GLfloat TextureVertex::getindex()
+int TextureVertex::getindex()
 {
 	return index;
 }
